backend bank deposit option added

+++ includes/class-ticket-booking-stripe.php
@@ -22,7 +22,7 @@ class Ticket_Booking_Stripe {
 
 		// Securely fetch all POST data
 		$payment_method = sanitize_text_field( $_POST['payment_method'] );
-		$total_amount   = intval( $_POST['total_amount'] ); // Amount in cents
+		$total_amount   = floatval( $_POST['total_amount'] ); // Amount in cents
 		$table_number   = sanitize_text_field( $_POST['table_number'] );
 		$seat_quantity  = intval( $_POST['seat_quantity'] );
 		$table_type     = sanitize_text_field( $_POST['table_type'] );
@@ -182,7 +182,7 @@ class Ticket_Booking_Stripe {
 		$pay_method    = sanitize_text_field( $_POST['payMethod'] );
 		$location      = sanitize_text_field( $_POST['location'] );
 		$bin_number    = sanitize_text_field( $_POST['bin_number'] );
-		$total_vat     = intval( $_POST['total_vat'] );
+		$total_vat     = floatval( $_POST['total_vat'] );
 		$vatPercentage = sanitize_text_field( $_POST['vatPercentage'] );
 
 		//update ticket details table
